getting the resize of the window, wip
