+++ components/LessonSection/NavDots/LiNavDotMobile.js
@@ -22,30 +22,6 @@ const LiNavDotMobile = ({ section, handleDotClick, index, setSectionDots }) => {
 
     const iconStyles = useMemo(() => getIconStyles(isInView, sectionId), [isInView, sectionId]);
 
-    const handleDocumentClick = event => {
-        const wasANavDotElementClicked = NAV_CLASSNAMES.some(className => event.target.classList.contains(className))
-
-        !wasANavDotElementClicked && setSectionDots(sectionDots => {
-            if (sectionDots?.length) {
-                return sectionDots.map(sectionDot => {
-                    return {
-                        ...sectionDot,
-                        willShowTitle: false,
-                    };
-                })
-            }
-
-            return sectionDots;
-        })
-    }
-
-    useEffect(() => {
-
-        document.body.addEventListener('click', handleDocumentClick);
-
-        return () => document.body.removeEventListener('click', handleDocumentClick);
-    }, [])
-
     return (
         <li
             key={index}
@@ -57,7 +33,7 @@ const LiNavDotMobile = ({ section, handleDotClick, index, setSectionDots }) => {
                 onClick={_ => handleDotClick(sectionId, true)}
                 className='sectionNavDot'
                 style={iconStyles} />
-            <div style={{ opacity: willShowTitle ? 1 : 0, width: 'auto', right: '18px', pointerEvents: 'none' }} className='position-absolute d-flex'>
+            <div style={{ opacity: willShowTitle ? 1 : 0, width: 'auto', right: '18px', pointerEvents: 'none', transition: "all .15s ease-in", transitionProperty: 'opacity' }} className='position-absolute d-flex'>
                 <span style={{ transition: "all .15s ease-in", backgroundColor: backgroundColor, border: '#363636 1px solid', transitionProperty: 'background-color, opacity, border' }} className='text-nowrap p-1 rounded'>{title}</span>
                 <span style={{ width: 55 }} className='d-flex d-md-none justify-content-center align-items-center ps-1 sectionTitleSpan'>
                     <span className="dotLine" />
