+++ pages/lessons/[loc]/index.js
@@ -72,6 +72,29 @@ const LessonDetails = ({ lesson, availLocs }) => {
 
   const [sectionDots, setSectionDots] = useState(getSectionDotsDefaultVal())
 
+  const handleDocumentClick = event => {
+    const wasANavDotElementClicked = NAV_CLASSNAMES.some(className => event.target.classList.contains(className))
+
+    !wasANavDotElementClicked && setSectionDots(sectionDots => {
+      if (sectionDots?.length) {
+        return sectionDots.map(sectionDot => {
+          return {
+            ...sectionDot,
+            willShowTitle: false,
+          };
+        })
+      }
+
+      return sectionDots;
+    })
+  }
+
+  useEffect(() => {
+    document.body.addEventListener('click', handleDocumentClick);
+
+    return () => document.body.removeEventListener('click', handleDocumentClick);
+  }, [])
+
   useEffect(() => {
     if (inView) {
       setSectionDots(sectionDots => {
@@ -189,7 +212,7 @@ export const getStaticProps = async ({ params: { loc: id } }) => {
   const lessons = await res.json();
   const lesson = lessons.find(lesson => `${lesson.id}` === `${id}`);
   const availLocs = lessons.filter(lesson => `${lesson.id}` === `${id}`).map((lesson) => lesson.locale);
-  
+
   if (!lesson?.Section?.procedure?.Data) {
     return { props: { lesson, availLocs } };
   }
